+++ src/images.c
@@ -58,7 +58,7 @@ void show_image_grayscale(struct image const *image) {
 
 void show_image_grayscale_at(struct image const *image, struct point pos) {
     for (size_t y = 0; y < image->h; y++) {
-        printf("\033[%u;%uH", pos.y + (uint32_t)y, pos.x);
+        printf("\033[%u;%uH", (uint32_t)pos.y + (uint32_t)y, (uint32_t)pos.x);
 
         for (size_t x = 0; x < image->w; x++) {
             uint8_t pixel = color_to_grayscale(image->pixels[y * image->w + x]);
@@ -67,7 +67,7 @@ void show_image_grayscale_at(struct image const *image, struct point pos) {
         printf("\n");
     }
 
-    printf("\033[%lu;%luH", pos.y + image->h, (size_t)1);
+    printf("\033[%lu;%luH", (uint32_t)pos.y + image->h, (size_t)1);
 }
 
 void image_destroy(struct image *image) {
